more gifs and love

+++ src/commands/say.js
@@ -7,8 +7,9 @@ class Say extends Command {
 
     on(message) {
             let messageToPrint = message.content.slice(8);
-            if (messageToPrint.length > 0) {
-                message.channel.send(messageToPrint);
+            
+        if (messageToPrint.length > 0 && message.author != "736342162886623313") {
+            message.channel.send(messageToPrint);
                 message.delete();
             } else {
                 message.channel.send("porfavor escribe un mensaje y no lo dejes en blanco nwn.");
